% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{add_proc_format_labels}
\alias{add_proc_format_labels}
\title{Add labels from proc_format file to SAS data file}
\usage{
add_proc_format_labels(df, proc_format, ...)
}
\arguments{
\item{df}{Input data frame read from \link[haven:read_sas]{haven::read_sas}}

\item{proc_format}{Either path to SAS file with \code{proc format} statements or
the results of \link{read_proc_format}.}

\item{...}{Arguments passed on to \code{read_proc_format}
\describe{
  \item{file}{Path to proc format file.}
  \item{verbose}{Writes out file path being read. May be removed in the future.}
  \item{missing_values}{Possible SAS values that should be treated as missing.}
}}
}
\description{
Add labels from proc_format file to SAS data file
}
\examples{
bdat <- haven::read_sas("/Volumes/Lab_Gerke/PLCO/Free PSA/freepsa_data_feb16_d080516.sas7bdat")
bdat2 <- add_proc_format_labels(bdat, "/Volumes/Lab_Gerke/PLCO/Free PSA/freepsa.sas_formats.feb16.d080516.sas")

}
