% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_diff.R
\name{align_data_frames}
\alias{align_data_frames}
\title{Align tibbles or data frames using grouping vars}
\usage{
align_data_frames(x, y, group_vars = NULL)
}
\description{
Uses \code{group_vars} to find missing rows in \code{x} or \code{y} and align final rows.
}
