% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-files.R
\name{auto_extract}
\alias{auto_extract}
\title{Auto Extract Zipped Files}
\usage{
auto_extract(path, ..., gunzip_args = NULL)
}
\arguments{
\item{path}{Path to a single zipped file or a folder containing files to be
unzipped.}

\item{...}{Arguments passed on to \link{untar} or \link{unzip} (or both for
\code{list} and \code{extdir}). The \code{files} argument will be ignored.}
}
\description{
Auto Extract Zipped Files
}
