% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_themes.R
\name{doc_new}
\alias{doc_new}
\title{Create New Document}
\usage{
doc_new(path, type = c("doc", "slides"))
}
\arguments{
\item{path}{Path to the location of the new document}

\item{type}{Type of document to create}
}
\description{
Creates a new R Markdown document of the requested type. There are currently
three templates, one for reports where the default is HTML based on the HTML
vignette template, and another with a Moffitt-styled \pkg{xaringan} theme. In
all cases, the document and supporting files are added to a directory with
the name given by the file.
}
\examples{
\dontrun{
doc_new("my_report.Rmd", "doc")
doc_new("my_slides.Rmd", "slides")
}

}
