% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2_utils.R
\name{ggsave_and_print}
\alias{ggsave_and_print}
\title{Save and/or Print Plot}
\usage{
ggsave_and_print(plot, filename, width = 20, height = 15,
  device = fs::path_ext(filename),
  include = knitr::opts_current$get("include") \%||\% TRUE, ...,
  force_save = FALSE, base_dir = getwd())
}
\arguments{
\item{plot}{Plot to save, defaults to last plot displayed.}

\item{filename}{File name to create on disk.}

\item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}

\item{device}{Device to use. Can be either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}

\item{include}{Mimics the \link{knitr} \code{include} chunk option. If \code{FALSE} then the
plot is saved but not printed. If \code{TRUE}, then the plot is saved and
printed.}

\item{...}{Other arguments passed on to graphics \code{device}.}

\item{force_save}{If \code{TRUE} the plot is saved, even if inside an interactive
session. The default is \code{FALSE}, meaning that plots are not saved when
in interactive mode.}

\item{base_dir}{Save figures into a specified directory}

\item{...}{Arguments passed on to \code{ggplot2::ggsave}
\describe{
  \item{filename}{File name to create on disk.}
  \item{plot}{Plot to save, defaults to last plot displayed.}
  \item{device}{Device to use. Can be either be a device function
(e.g. \code{\link[=png]{png()}}), or one of "eps", "ps", "tex" (pictex),
"pdf", "jpeg", "tiff", "png", "bmp", "svg" or "wmf" (windows only).}
  \item{path}{Path to save plot to (combined with filename).}
  \item{scale}{Multiplicative scaling factor.}
  \item{width}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
  \item{height}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
  \item{units}{Plot size in \code{units} ("in", "cm", or "mm").
If not supplied, uses the size of current graphics device.}
  \item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}
  \item{limitsize}{When \code{TRUE} (the default), \code{ggsave} will not
save images larger than 50x50 inches, to prevent the common error of
specifying dimensions in pixels.}
}}
}
\description{
A thin wrapper around \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}} that takes a plot as the first
argument and \code{filename} as the second (the reverse of \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}),
but that doesn't save the plot when called interactively. Works best inside R
Markdown documents, allowing you to view the figure when running the code
chunk interactively but ensure that a specific copy of the figure is saved to
a specific directory during the markdown rendering.
}
\examples{
\dontrun{
iris \%>\%
{
  ggplot(.) +
    aes(Sepal.Length, Sepal.Width, color = Species) +
    geom_point()
} \%>\%
ggsave_and_print("iris_plot.png", width = 10, height = 6)
}

}
