% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{grk_style}
\alias{grk_style}
\alias{grk_style_transformer}
\alias{use_grk_style}
\alias{grk_style_file}
\alias{grk_style_dir}
\alias{grk_style_pkg}
\title{The grkmisc Code Style}
\usage{
grk_style_transformer(...)

use_grk_style()

grk_style_file(path, ..., transformers = grk_style_transformer())

grk_style_dir(path, ..., transformers = grk_style_transformer())

grk_style_pkg(pkg, ..., transformers = grk_style_transformer())
}
\arguments{
\item{...}{Arguments passed to underling \pkg{styler} functions (identified
by removing the \code{grk_} prefix)}

\item{path}{A character vector with paths to files to style.}

\item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}

\item{pkg}{Path to a (subdirectory of an) R package.}
}
\description{
Use \code{\link[styler:style_text]{styler::style_text()}} to format code according to the unofficial
\pkg{grkmisc} style guide.
}
\section{Functions}{
\itemize{
\item \code{grk_style_transformer}: A code transformer for use with \code{\link[styler:style_text]{styler::style_text()}}

\item \code{use_grk_style}: Set the \pkg{grkmisc} style as the default style for
\pkg{styler} addins.

\item \code{grk_style_file}: Style a file using the \pkg{grkmisc} code style

\item \code{grk_style_dir}: Style a directory using the \pkg{grkmisc} code style

\item \code{grk_style_pkg}: Style a package using the \pkg{grkmisc} code style
}}

\examples{
\dontrun{
use_grk_style()
# Use styler addins
styler:::style_selection()
}

grk_style_text(
  "x = paste(\\n'a', 'b')",
  transformers = grk_style_transformer()
)

}
