% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grkmisc-package.R
\docType{package}
\name{grkmisc-package}
\alias{grkmisc}
\alias{grkmisc-package}
\title{grkmisc: A package of miscellaneously helpful functions}
\description{
A package of miscellaneously helpful
    functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gerkelab/grkmisc}
  \item Report bugs at \url{https://github.com/gerkelab/grkmisc/issues}
}

}
\author{
\strong{Maintainer}: Garrick Aden-Buie \email{Garrick.Aden-Buie@moffitt.org}

}
\keyword{internal}
