% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_na.R
\name{if_na.data.frame}
\alias{if_na.data.frame}
\title{Replace Missing Values in Data Frames}
\usage{
\method{if_na}{data.frame}(.x, ..., then = NULL, otherwise = NULL, vars = NULL)
}
\arguments{
\item{.x}{Input vector of values}

\item{...}{Optional named arguments specifying column-wise replacement values
for missing values. Other arguments are ignored by \code{if_na()} when used.}

\item{then}{Global replacement value to be used in place of the default.}

\item{otherwise}{Ignored.}

\item{vars}{Optional vector of column names that should receive the missing
value replacement. Use the \code{vars()} helper to provide bare column names.}
}
\description{
Replaces missing values in data frames, similar to \code{\link[tidyr:replace_na]{tidyr::replace_na()}} with
a few extras. There are several ways to use this function. If called without
any arguments, all missing values will be replaced by the default type.
You can change the global default value with the \code{then} argument, and you
can subset the columns to be affected with the \code{vars} argument. Finally, you
can specify column-wise replacement values as named arguments, in which case
the other arguments are ignored.
}
\examples{
df <- dplyr::tibble(
  x = c(1, 2, NA), y = c("a", NA, "b"), z = list(1:5, NULL, 10:20)
)

df \%>\% if_na()

df \%>\% if_na(then = "0")

df \%>\% if_na(vars = dplyr::vars(x, y))

df \%>\% if_na(x = 3, y = "c")

}
