% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/if_na.R
\name{if_na}
\alias{if_na}
\title{Replace Missing Values}
\usage{
if_na(.x, ..., then = default_value(x), otherwise = NULL)
}
\arguments{
\item{...}{Additional expressions to condition replacement. NA values are
only replaced when the additional expression matches.}

\item{then}{Replacement value. Defaults to a value based on the input type
and favors \code{0}, \code{FALSE}, \code{""} or empty lists according to the input type.}

\item{otherwise}{If given, provides a secondary replacement value to be
subsituted for missing values when the conditions in \code{...} are \code{FALSE}.}

\item{x}{Input vector of values}
}
\description{
Replaces missing values with the given value, or optionally conditions the
replacement on the expressions given in \code{...}, which are applied to the
inputs in the same way as \code{filter}.
}
\examples{
library(dplyr)
storms <- storms \%>\%
  filter(name == "Erika", year == 2003) \%>\%
  select(name, year, status, wind, contains("diameter"))

# Default is replacement with appropriate default (0, FALSE, "", etc.)
storms \%>\%
  mutate(hu_diameter = if_na(hu_diameter))

# Replacement can be single value or vector of values
storms \%>\%
  mutate(
    ts_diameter = if_na(ts_diameter, then = 30),
    hu_diameter = if_na(hu_diameter, then = rnorm(11, 100))
  )

# Replacement can be predicated on other conditions in addition to missingness
storms \%>\%
  mutate(
    hu_diameter = if_na(hu_diameter, then = 30, status == "hurricane")
  )

# Can provide a secondary value for missing values when the conditions
# are not met
storms \%>\%
  mutate(
    ts_diameter = if_na(ts_diameter, then = wind, status == "tropical storm", otherwise = 0)
  )

}
