% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{insert_quoted_equals}
\alias{insert_quoted_equals}
\title{Fixes issue where an equals character might be inside double quotes, in
which case when splitting on "=" there may be an extra (or multiple splits).
This function puts equals back inside \emph{double quotes}. Doesn't work with
single quotes.}
\usage{
insert_quoted_equals(x)
}
\description{
Fixes issue where an equals character might be inside double quotes, in
which case when splitting on "=" there may be an extra (or multiple splits).
This function puts equals back inside \emph{double quotes}. Doesn't work with
single quotes.
}
