% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insert_styled_text}
\alias{insert_styled_text}
\title{Insert Styled Text}
\usage{
insert_styled_text(x = NULL, ...)
}
\arguments{
\item{x}{The code as a character string}

\item{...}{Arguments passed on to \code{style_console}
\describe{
  \item{x}{Text to style}
}}
}
\description{
Inserts text from clipboard at cursor position, styled according to the
\href{http://style.tidyverse.org/}{tidyverse style guide}. Powers the RStudio
addin.
}
