% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knitr_bold_row}
\alias{knitr_bold_row}
\title{Embolden rows matching an expression}
\usage{
knitr_bold_row(x, ..., cols = NULL, format = c("markdown", "html",
  "latex"))
}
\arguments{
\item{x}{A data frame}

\item{...}{The filter expressions}

\item{cols}{Columns that should be emphasized, default \code{NULL} or all columns}

\item{format}{One of \code{"markdown"}, \code{"html"}, or \code{"latex"}}
}
\description{
Emboldens rows matching an expression, with optionally selected columns. For
use with \code{\link[knitr:kable]{knitr::kable()}}.
}
\examples{
x <- dplyr::starwars[1:4, 1:5]
knitr_bold_row(x, height < 100)

knitr_bold_row(x, mass <= 75, height <= 170, cols = c("name", "height"), format = "html")

x \%>\%
  knitr_bold_row(height > 170, cols = "hair_color") \%>\%
  knitr::kable()

}
