% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knitr_list_cache}
\alias{knitr_list_cache}
\title{List cached knitr chunks}
\usage{
knitr_list_cache(cache_path = NULL, names_only = TRUE)
}
\arguments{
\item{cache_path}{Path to cache directory}

\item{names_only}{Print only names of cached knitr chunks? Otherwise returns
full absolute path names for loading the chunks.}
}
\description{
List cached knitr chunks
}
\seealso{
Other knitr helpers: 
\code{\link{knitr_load_cache}()}
}
\concept{knitr helpers}
