% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr.R
\name{knitr_load_cache}
\alias{knitr_load_cache}
\title{Lazy load results from a knitr cache}
\usage{
knitr_load_cache(chunk = NULL, cache_path = NULL)
}
\arguments{
\item{chunk}{Name or names of chunk to load}

\item{cache_path}{Path to cache directory}
}
\description{
Lazy load results from a knitr cache
}
\seealso{
Other knitr helpers: 
\code{\link{knitr_list_cache}()}
}
\concept{knitr helpers}
