% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{logger}
\alias{logger}
\title{Pipe-Capable Log Messages}
\usage{
logger(.data = NULL, msg = "", level = "INFO")
}
\arguments{
\item{.data}{This argument is returned as-is.}

\item{msg}{The log message, processed using \link[glue:glue]{glue::glue}. You can refernce
the data passed in via \code{.data}.}

\item{level}{The log level, default is "INFO" and possible values include
"DEBUG", "INFO", "WARNING", "ERROR", and "FATAL".}
}
\description{
Writes a log message, suitable for use inside pipes, meaning that the first
argument is passed through untouched. The log message is processed using
\link[glue:glue]{glue::glue}, and you can reference the incoming data as \code{.data} inside the
log message.
}
\section{Options}{


You can set the maximum debug level and the log output locations using the
global options \code{grkmisc.log_level} and \code{grkmisc.log_output}. The defaults
are \code{"INFO"} and \code{stdout()} respectively.
}

\examples{
library(dplyr)
iris \%>\%
  logger("Starting with {nrow(.data)} rows") \%>\%
  filter(Species == "setosa") \%>\%
  logger("Filtered to {nrow(.data)} rows and {ncol(.data)} columns") \%>\%
  head()

}
