% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grkmisc-package.R
\docType{data}
\name{moffitt_colors}
\alias{moffitt_colors}
\title{Moffitt Colors}
\format{A named list of 7 colors: blue, light_blue, green, orange,
grey, yellow, and red.}
\usage{
moffitt_colors
}
\description{
Color codes for the colors listed in the Moffitt Brand Guide 2014.
}
\section{CSS}{

A CSS palette is included as well, that can be found via
\code{system.file("moffitt-colors.css", package = "grkmisc")}.
}

\keyword{datasets}
