% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pretty_string}
\alias{pretty_string}
\alias{format_pretty_string}
\title{Truncate and wrap strings}
\usage{
pretty_string(
  x,
  truncate_at = 80,
  truncate_with = "...",
  trim = TRUE,
  wrap_at = 40
)

format_pretty_string(
  truncate_at = 80,
  truncate_with = "...",
  trim = TRUE,
  wrap_at = 40
)
}
\arguments{
\item{x}{Strings}

\item{truncate_at}{Maximum total string length prior to wrapping. Default is
80 characters. Use \code{NULL} to skip truncation.}

\item{truncate_with}{Character string that is added at the end of each string
to indicate that the string was truncated. Eats into string length. Default
is \code{"..."}; set to \code{NULL} or \code{""} to skip.}

\item{trim}{Should whitespace be trimmed? Default is \code{TRUE}.}

\item{wrap_at}{Wraps string at given length, passed to \code{\link[stringr:str_wrap]{stringr::str_wrap()}}.}
}
\description{
Truncates, trims, and wraps strings. Built for ggplot2 plots with long
string labels.
}
\section{Functions}{
\itemize{
\item \code{format_pretty_string}: Provides a pretty string formatter for ggplot2 labels
}}

\examples{
text <- "Lorem ipsum dolor sit amet, consectetur adipiscing elit"
pretty_string(text, truncate_at = 20, wrap_at = NULL)
pretty_string(text, truncate_at = NULL, wrap_at = 10)

library(ggplot2)
set.seed(654321)
ex <- dplyr::data_frame(
  label = sample(stringr::sentences, 3),
  value = runif(3, 0, 10)
)

g <- ggplot(ex) +
  aes(value, label) +
  geom_point()

g

g + scale_y_discrete(
  label = format_pretty_string(truncate_at = 25)
)

g + scale_y_discrete(
  label = format_pretty_string(truncate_at = NULL, wrap_at = 20)
)
}
