% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{read_proc_format}
\alias{read_proc_format}
\title{Read Proc Format File}
\usage{
read_proc_format(file, verbose = FALSE, missing_values = paste0(".",
  c("", 1:9, LETTERS)))
}
\arguments{
\item{file}{Path to proc format file.}

\item{verbose}{Writes out file path being read. May be removed in the future.}

\item{missing_values}{Possible SAS values that should be treated as missing.}
}
\description{
Read Proc Format File
}
\examples{
\dontrun{
read_proc_format("freepsa.sas_formats.feb16.d080516.sas")
}

}
\seealso{
Other SAS helper functions: \code{\link{add_proc_format_labels}},
  \code{\link{read_sas_with_format}}
}
\concept{SAS helper functions}
