% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{read_sas_with_format}
\alias{read_sas_with_format}
\title{Read SAS file and label according to PROC FORMAT}
\usage{
read_sas_with_format(file_bdat, file_format, verbose = FALSE,
  missing_values = paste0(".", c("", 1:9, LETTERS)), ...)
}
\arguments{
\item{file_bdat}{Path to the \code{sas7bdat} file to be read by \link[haven:read_sas]{haven::read_sas}}

\item{file_format}{Path to the SAS formats file to be read by \link{read_proc_format}}

\item{verbose}{Writes out file path being read. May be removed in the future.}

\item{missing_values}{Possible SAS values that should be treated as missing.}

\item{...}{Arguments passed on to \code{haven::read_sas}
\describe{
  \item{data_file}{Path to data and catalog files. The files are
processed with \code{\link[readr:datasource]{readr::datasource()}}.}
  \item{catalog_file}{Path to data and catalog files. The files are
processed with \code{\link[readr:datasource]{readr::datasource()}}.}
  \item{encoding}{The character encoding used for the
\code{data_file} and \code{catalog_encoding} respectively. A value of \code{NULL}
uses the encoding specified in the file; use this argument to override it
if it is incorrect.}
  \item{catalog_encoding}{The character encoding used for the
\code{data_file} and \code{catalog_encoding} respectively. A value of \code{NULL}
uses the encoding specified in the file; use this argument to override it
if it is incorrect.}
  \item{cols_only}{A character vector giving an experimental way to read in
only specified columns.}
}}
}
\description{
Reads a SAS \code{sas7bdat} file using \link[haven:read_sas]{haven::read_sas} and then applies the
labels given in the formats file (containing \code{PROC FORMAT} statements).
The labels use \link[labelled:labelled]{labelled::labelled} and may be coerced to factors if desired.
}
\seealso{
Other SAS helper functions: \code{\link{add_proc_format_labels}},
  \code{\link{read_proc_format}}
}
\concept{SAS helper functions}
