% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{scale_moffitt}
\alias{scale_moffitt}
\alias{scale_colour_moffitt}
\alias{scale_color_moffitt}
\alias{scale_fill_moffitt}
\title{Moffitt Color Scales for ggplot2}
\usage{
scale_colour_moffitt(color_other = "grey", direction = 1, ...)

scale_color_moffitt(color_other = "grey", direction = 1, ...)

scale_fill_moffitt(color_other = "grey", direction = 1, ...)
}
\arguments{
\item{color_other}{When the data contains two values, the second value takes
this color. Can be any of the colors in \link{moffitt_colors} other than blue:
green, red, orange, light_blue, yellow, or grey (default).}

\item{direction}{Reverses the direction of the color scale when \code{direction}
is less than 0, i.e. -1.}

\item{...}{Arguments passed on to \code{ggplot2::discrete_scale}
\describe{
  \item{aesthetics}{The names of the aesthetics that this scale works with}
  \item{scale_name}{The name of the scale}
  \item{palette}{A palette function that when called with a single integer
argument (the number of levels in the scale) returns the values that
they should take}
  \item{name}{The name of the scale. Used as axis or legend title. If
\code{waiver()}, the default, the name of the scale is taken from the first
mapping used for that aesthetic. If \code{NULL}, the legend title will be
omitted.}
  \item{breaks}{One of:
\itemize{
\item \code{NULL} for no breaks
\item \code{waiver()} for the default breaks computed by the
transformation object
\item A character vector of breaks
\item A function that takes the limits as input and returns breaks
as output
}}
  \item{labels}{One of:
\itemize{
\item \code{NULL} for no labels
\item \code{waiver()} for the default labels computed by the
transformation object
\item A character vector giving labels (must be same length as \code{breaks})
\item A function that takes the breaks as input and returns labels
as output
}}
  \item{limits}{A character vector that defines possible values of the scale
and their order.}
  \item{expand}{Vector of range expansion constants used to add some
padding around the data, to ensure that they are placed some distance
away from the axes. Use the convenience function \code{\link[=expand_scale]{expand_scale()}}
to generate the values for the \code{expand} argument. The defaults are to
expand the scale by 5\% on each side for continuous variables, and by
0.6 units on each side for discrete variables.}
  \item{na.translate}{Unlike continuous scales, discrete scales can easily show
missing values, and do so by default. If you want to remove missing values
from a discrete scale, specify \code{na.translate = FALSE}.}
  \item{na.value}{If \code{na.translate = TRUE}, what value aesthetic
value should missing be displayed as? Does not apply to position scales
where \code{NA} is always placed at the far right.}
  \item{drop}{Should unused factor levels be omitted from the scale?
The default, \code{TRUE}, uses the levels that appear in the data;
\code{FALSE} uses all the levels in the factor.}
  \item{guide}{A function used to create a guide or its name. See
\code{\link[=guides]{guides()}} for more info.}
  \item{position}{The position of the axis. "left" or "right" for vertical
scales, "top" or "bottom" for horizontal scales}
  \item{super}{The super class to use for the constructed scale}
}}
}
\description{
Color scales based on the Moffitt Branding Guidelines, 2014.
}
\examples{
library(ggplot2)
ggplot(mtcars) +
  aes(mpg, wt, color = paste(vs)) +
  geom_point(size = 2) +
  theme_moffitt() +
  scale_color_moffitt()

ggplot(mtcars) +
  aes(mpg, wt, color = paste(vs)) +
  geom_point(size = 2) +
  theme_moffitt() +
  scale_color_moffitt(color_other = "green")

ggplot(mtcars) +
  aes(mpg, wt, color = paste(carb)) +
  geom_point(size = 2) +
  theme_moffitt() +
  scale_color_moffitt()

dplyr::count(mpg, class, sort = TRUE) \%>\%
  dplyr::mutate(class = factor(class, levels = class)) \%>\%
  ggplot() +
  aes(class, n, fill = class) +
  geom_col() +
  coord_flip() +
  theme_moffitt() +
  scale_fill_moffitt()

}
\seealso{
\link{moffitt_colors} \code{\link[=theme_moffitt]{theme_moffitt()}}
}
