% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{shorten_github_url}
\alias{shorten_github_url}
\title{Create Shortlink for a GitHub URL using git.io}
\usage{
shorten_github_url(url, vanity_code = NULL)
}
\arguments{
\item{url}{The GitHub URL to convert to a shortlink}

\item{vanity_code}{Request a specific shortcode instead of the randomly
generated name. Helpful for easily sharing a link verbally.}
}
\value{
A (hopefully) shorter URL
}
\description{
Create Shortlink for a GitHub URL using git.io
}
