% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{style_console}
\alias{style_console}
\title{Style text from console}
\usage{
style_console(x, ...)
}
\arguments{
\item{x}{Text to style}

\item{...}{Arguments passed on to \code{styler::style_text}
\describe{
  \item{text}{A character vector with text to style.}
  \item{style}{A function that creates a style guide to use, by default
\code{\link[=tidyverse_style]{tidyverse_style()}} (without the parentheses). Not used
further except to construct the argument \code{transformers}. See
\code{\link[=style_guides]{style_guides()}} for details.}
  \item{transformers}{A set of transformer functions. This argument is most
conveniently constructed via the \code{style} argument and \code{...}. See
'Examples'.}
}}
}
\description{
Style text from console
}
