% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_moffitt}
\alias{theme_moffitt}
\alias{theme_grk}
\title{A Clean Theme with Moffitt Style}
\usage{
theme_moffitt(base_family = "Overpass",
  axis_text_family = "Overpass Mono", axis_title_family = base_family,
  axis_title_bold = FALSE, axis_title_just = "cc",
  axis_title_size = 13, subtitle_size = 13,
  default_geom_font = "Fira Sans Condensed",
  default_geom_color = grkmisc::moffitt_colors$blue,
  axis_text_color = "#6e6e6e", plot_caption_color = axis_text_color,
  panel_border_color = axis_text_color, ..., use_showtext = TRUE)

theme_grk(base_family = "Fira Sans",
  axis_text_family = paste(sub(" ?(Mono|Sans|Serif)", "", base_family),
  "Mono"), axis_title_family = base_family, axis_title_bold = FALSE,
  axis_title_just = "cc", axis_title_size = 13, subtitle_size = 13,
  default_geom_font = "Fira Sans Condensed",
  default_geom_color = grkmisc::moffitt_colors$blue,
  axis_text_color = "#6e6e6e", plot_caption_color = axis_text_color,
  panel_border_color = axis_text_color, ..., use_showtext = TRUE)
}
\arguments{
\item{base_family}{base font family and size}

\item{axis_text_family}{The font family for axis ticks text labels.
Passed to \code{family} in \code{axis.text} in \code{\link[ggplot2:theme]{ggplot2::theme()}}.}

\item{axis_title_family}{axis title font family, face and size}

\item{axis_title_bold}{If \code{TRUE}, the axis title's will be bold.}

\item{axis_title_just}{axis title font justification, one of \code{[blmcrt]}}

\item{axis_title_size}{axis title font family, face and size}

\item{subtitle_size}{plot subtitle family, face and size}

\item{default_geom_font}{Change the default ggplot2 geom fonts to the
specified font. The default is "Fira Sans Condensed", which tends to look
good in constrained space.}

\item{axis_text_color}{Color of axis text}

\item{plot_caption_color}{Color of the plot caption text}

\item{panel_border_color}{Color of the panel border}

\item{...}{Arguments passed on to \code{hrbrthemes::theme_ipsum}
\describe{
  \item{base_family}{base font family and size}
  \item{base_size}{base font family and size}
  \item{plot_title_family}{plot title family, face, size and margi}
  \item{plot_title_face}{plot title family, face, size and margi}
  \item{plot_title_size}{plot title family, face, size and margi}
  \item{plot_title_margin}{plot title family, face, size and margi}
  \item{subtitle_family}{plot subtitle family, face and size}
  \item{subtitle_face}{plot subtitle family, face and size}
  \item{subtitle_size}{plot subtitle family, face and size}
  \item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}
  \item{strip_text_family}{facet label font family, face and size}
  \item{strip_text_face}{facet label font family, face and size}
  \item{strip_text_size}{facet label font family, face and size}
  \item{caption_family}{plot caption family, face, size and margin}
  \item{caption_face}{plot caption family, face, size and margin}
  \item{caption_size}{plot caption family, face, size and margin}
  \item{caption_margin}{plot caption family, face, size and margin}
  \item{axis_text_size}{font size of axis text}
  \item{axis_title_family}{axis title font family, face and size}
  \item{axis_title_face}{axis title font family, face and size}
  \item{axis_title_size}{axis title font family, face and size}
  \item{axis_title_just}{axis title font justification, one of \code{[blmcrt]}}
  \item{plot_margin}{plot margin (specify with \code{\link[ggplot2:margin]{ggplot2::margin()}})}
  \item{grid_col}{grid & axis colors; both default to \code{#cccccc}}
  \item{axis_col}{grid & axis colors; both default to \code{#cccccc}}
  \item{grid}{panel grid (\code{TRUE}, \code{FALSE}, or a combination of \code{X}, \code{x}, \code{Y}, \code{y})}
  \item{axis}{add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}"}
  \item{ticks}{ticks if \code{TRUE} add ticks}
}}

\item{use_showtext}{Should \link{showtext} and \link{sysfonts} be used to register
font families from Google? Default is \code{TRUE}.}

\item{default_color}{Changes default colors of bars and points to
the value given. Set to \code{NULL} to avoid changing these colors. Note that
these colors are used only when color or fill is not mapped to the data.}
}
\description{
A clean theme based on \code{\link[hrbrthemes:theme_ispum]{hrbrthemes::theme_ispum()}} that uses
\href{https://fonts.google.com/specimen/Overpass}{Overpass}, a free Google Font
that resembles Freeway, the (proprietary) font chosen by MCC branding. This
theme works best when set globally using \code{\link[ggplot2:theme_set]{ggplot2::theme_set()}} as it will
automatically download and register the correct fonts from Google Font using
\link{sysfonts} and \link{showtext} (if installed) and will change the default
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}} related fonts as well. For an alternative look that I
prefer, use \code{theme_grk()}, which simply uses "Fira Sans" as the base font.
}
\examples{
\dontrun{
theme_set(theme_moffitt())
theme_set(theme_grk())
# Set base theme without changing geom defaults
theme_set(theme_moffitt(default_geom_font = NULL, default_geom_color = NULL))
}

library(ggplot2)
g <- ggplot(mtcars, aes(mpg, wt)) +
  geom_point() +
  geom_text(aes(label = cyl), vjust = 0.5, hjust = 1.65) +
  labs(x = "Fuel efficiency (mpg)",
       y = "Weight (tons)",
       title = "Seminal ggplot2 scatterplot example",
       subtitle = "A plot that is only useful for demonstration purposes",
       caption = "Brought to you by the letter 'g'")
g + theme_moffitt()
g + theme_moffitt(default_geom_color = moffitt_colors$green)
g + theme_grk()

}
