% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_diff.R
\name{tidy_diff}
\alias{tidy_diff}
\title{Compare Two Data.Frames}
\usage{
tidy_diff(x, y, ignore = NULL, group_vars = ignore, align = FALSE)
}
\arguments{
\item{x}{\code{<tbl|df>} Reference data frame}

\item{y}{\code{<tbl|df>} Comparison data frame}

\item{ignore}{\code{<chr:NULL>} ID column(s) that are ignored from the differencing}

\item{group_vars}{\code{<chr:ignore>} Column names of grouping variables for alignment,
takes default from \code{ignore}.}

\item{align}{\code{<lgl:FALSE>} Should alignmment be performed based on the
grouping variables or the \code{group_vars} parameter?}
}
\value{
Object of class \code{tidy_diff} that can be printed via \code{print(obj)} or
plotted with \link{ggplot2} via \code{plot(obj)}.
}
\description{
Finds row-wise differences in two data frames that are assumed to be
pre-arranged.
}
