% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_grk_gitignore}
\alias{use_grk_gitignore}
\title{Use Default .gitignore File}
\usage{
use_grk_gitignore(directory = here::here(), browse = TRUE,
  overwrite = yesno::yesno("Overwrite .gitignore? Select no to append to existing .gitignore file"))
}
\arguments{
\item{directory}{Directory for the .gitignore, defaults to project root.}

\item{browse}{Open the .gitignore file after modifying?}

\item{overwrite}{Should existing file be overwritten?}
}
\description{
Copies the grkmisc default .gitignore file into the provided directory.
}
