% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_grk_starter_project}
\alias{use_grk_starter_project}
\title{Create a default project skeleton}
\usage{
use_grk_starter_project(path)
}
\arguments{
\item{path}{Path where the project will be started}
}
\description{
Uses \link{usethis} to create an empty project, with an RStudio project file and
an initialized git repository in the specified path.
}
