% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{use_starter_package}
\alias{use_starter_package}
\title{Create a default package skeleton}
\usage{
use_starter_package(path, github = TRUE, github_org = NULL,
  github_private = FALSE, open = TRUE,
  title = "What the Package Does (One Line, Title Case)",
  description = "What the package does (one paragraph)")
}
\arguments{
\item{path}{Path (and thus package name) where the package will be stored}

\item{github}{Should a github repository be created?}

\item{github_org}{The organization where the repo should be created. If
unspecified (or \code{NULL}), the default is to create the package in your
personal account.}

\item{github_private}{Should the GitHub repo be private?}

\item{open}{Should the package be opened? If \code{TRUE}, tries to open in RStudio
and falls back to changing the working directory.}

\item{title}{Package title: "What the Package Does (One Line, Title Case)"}

\item{description}{Package description: "What the package does (one
paragraph)"}
}
\description{
Create a default package skeleton
}
\section{Default Package Description}{

See \code{\link[usethis:use_description]{usethis::use_description()}} for information about setting your default
package DESCRIPTION file fields. The help file there provides an example of
how to set \code{"usethis.description"} in your \code{.Rprofile}.
}

