% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{grk_style}
\alias{grk_style}
\alias{use_grk_style}
\alias{grk_style_transformer}
\alias{grk_style_text}
\alias{grk_style_file}
\alias{grk_style_dir}
\alias{grk_style_pkg}
\title{The grkmisc Code Style}
\usage{
use_grk_style()

grk_style_transformer(...)

grk_style_text(text, ...)

grk_style_file(path, ...)

grk_style_dir(path, ...)

grk_style_pkg(pkg, ...)
}
\arguments{
\item{...}{Arguments passed to underling \pkg{styler} functions (identified
by removing the \code{grk_} prefix), except for \code{transofrmers}, which is set to
the \code{grk_style_transformer()} internally.}

\item{text}{A character vector with text to style.}

\item{path}{A character vector with paths to files to style.}

\item{pkg}{Path to a (subdirectory of an) R package.}
}
\description{
Use \code{\link[styler:style_text]{styler::style_text()}} to format code according to the unofficial
\pkg{grkmisc} style guide.
}
\section{Functions}{
\itemize{
\item \code{use_grk_style}: Set the \pkg{grkmisc} style as the default style for
\pkg{styler} addins.

\item \code{grk_style_transformer}: A code transformer for use with \code{\link[styler:style_text]{styler::style_text()}}

\item \code{grk_style_text}: Style text using the \pkg{grkmisc} code style

\item \code{grk_style_file}: Style a file using the \pkg{grkmisc} code style

\item \code{grk_style_dir}: Style a directory using the \pkg{grkmisc} code style

\item \code{grk_style_pkg}: Style a package using the \pkg{grkmisc} code style
}}

\section{Using the grkmisc code style}{
 You can set the \pkg{grkmisc} code
style as the default code style for \pkg{styler} (and its associated
RStudio addins, like "Style active file" and "Style selection") by calling
\code{grkmisc::use_grk_style()}. If you would rather set this option globally
for all session, you can add the following to your \code{.Rprofile}:\preformatted{options(styler.addins_style_transformer = "grkmisc::grk_style_transformer()")
}
}

\examples{
\dontrun{
use_grk_style()
# Use styler addins
styler:::style_selection()
}

ex_code <- "mtcars \%>\% mutate(mpg = mpg * 2,\n\t\t     cyl = paste(cyl)) \%>\% head()"
cat(ex_code)

grk_style_text(ex_code)

}
