% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{if_na}
\alias{if_na}
\title{Replace NA values}
\usage{
if_na(x, value = default_value(x), ..., value_else = NULL)
}
\arguments{
\item{x}{Input vector of values}

\item{value}{Replacement value, the default value is based on the input
value type and favors \code{0}, \code{FALSE}, or empty strings, lists or data frames
according to the input type.}

\item{...}{Additional expressions to condition replacement. NA values are
only replaced when the additional expression matches.}

\item{value_else}{If given, then provides a secondary replacement value to
be subsituted for missing values when the conditions in \code{...} are \code{FALSE}.}
}
\description{
Replaces NA values with the given value, or optionally conditions the
replacement on the expressions given in \code{...}, which are applied to the
inputs in the same way as \code{filter}.
}
\examples{
library(dplyr)
storms <- storms \%>\%
  filter(name == "Erika", year == 2003) \%>\%
  select(name, year, status, wind, contains("diameter"))

# Default is replacement with appropriate default (0, FALSE, "", etc.)
storms \%>\%
  mutate(hu_diameter = if_na(hu_diameter))

# Replacement can be single value or vector of values
storms \%>\%
  mutate(
    ts_diameter = if_na(ts_diameter, 30),
    hu_diameter = if_na(hu_diameter, rnorm(11, 100))
  )

# Replacement can be predicated on other conditions in addition to missingness
storms \%>\%
  mutate(
    hu_diameter = if_na(hu_diameter, 30, status == "hurricane")
  )

# Can provide a secondary value for missing values when the conditions
# are not met
storms \%>\%
  mutate(
    ts_diameter = if_na(ts_diameter, wind, status == "tropical storm", value_else = 0)
  )


}
