% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{pretty_num}
\alias{pretty_num}
\title{Pretty print scaled numbers}
\usage{
pretty_num(x, units = c(k = 1000, M = 1e+06, B = 1e+09),
  decimal_digits = 1, ...)
}
\arguments{
\item{x}{Vector of numbers}

\item{units}{Named list with lower bounds for units. Names are units or unit
abbreviations.}

\item{decimal_digits}{Number of digits to show after the decimal point in the
selected unit.}
}
\description{
Pretty print scaled numbers
}
\examples{
pretty_num(1234)
pretty_num(12345, decimal_digits = 3)
pretty_num(123456789)
pretty_num(12345678900)
pretty_num(c(1234, 1234567, 12345678900))
}
