% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{read_proc_format}
\alias{read_proc_format}
\title{Read Proc Format File}
\usage{
read_proc_format(file)
}
\arguments{
\item{file}{Path to proc format file.}
}
\description{
Read Proc Format File
}
\examples{
read_proc_format("/Volumes/Lab_Gerke/PLCO/Free PSA/freepsa.sas_formats.feb16.d080516.sas")

}
