% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_proc_format.R
\name{read_proc_format}
\alias{read_proc_format}
\title{Read Proc Format File}
\usage{
read_proc_format(file, verbose = FALSE, missing_values = c(".F", ".M",
  ".N"))
}
\arguments{
\item{file}{Path to proc format file.}

\item{verbose}{Writes out file path being read. May be removed in the future.}

\item{missing_values}{Possible SAS values that should be treated as missing.}
}
\description{
Read Proc Format File
}
\examples{
read_proc_format("/Volumes/Lab_Gerke/PLCO/Free PSA/freepsa.sas_formats.feb16.d080516.sas")

}
