% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_diff.R
\name{tidy_diff}
\alias{tidy_diff}
\title{Compare Two Data.Frames}
\usage{
tidy_diff(x, y, .id = NULL)
}
\arguments{
\item{x}{<tbl|df> Reference data frame}

\item{y}{<tbl|df> Comparison data frame}

\item{.id}{<chr> ID column(s) that are ignored from the differencing}
}
\value{
Object of class \code{tidy_diff} that can be printed via \code{print(obj)} or
plotted with \link{ggplot2} via \code{plot(obj)}.
}
\description{
Finds row-wise differences in two data frames that are assumed to be
pre-arranged.
}
